+++ packages/navigator/src/browser/navigator-widget.tsx
@@ -116,7 +116,7 @@ export class FileNavigatorWidget extends FileTreeWidget {
 
     protected getContainerTreeNode(): TreeNode | undefined {
         const root = this.model.root;
-        if (this.workspaceService.isMultiRootWorkspaceOpened) {
+        if (this.workspaceService.isMultiRootWorkspaceEnabled) {
             return root;
         }
         if (WorkspaceNode.is(root)) {